+++ backend/onyx/agents/agent_search/dr/utils.py
@@ -13,7 +13,7 @@ from onyx.agents.agent_search.shared_graph_utils.operators import (
 from onyx.context.search.models import InferenceSection
 from onyx.context.search.models import SavedSearchDoc
 from onyx.context.search.utils import chunks_or_sections_to_search_docs
-from onyx.tools.tool_implementations.internet_search.internet_search_tool import (
+from onyx.tools.tool_implementations.web_search.web_search_tool import (
     WebSearchTool,
 )
